requirements.txt
# ROS2 Bag Annotator - Python Dependencies
# Install with: pip install -r requirements.txt

# Core ROS2 dependencies (usually provided by ROS2 installation)
# rclpy
# rosidl-runtime-py
# rosbag2-py

# GUI Framework
tkinter  # Usually comes with Python, but listed for completeness

# Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.20.0

# Computer Vision Bridge
# cv-bridge  # Provided by ROS2

# Data Handling
PyYAML>=5.4.0
sqlite3  # Built-in Python module

# Development and Testing
pytest>=6.0.0
pytest-mock>=3.0.0
pytest-cov>=2.0.0

# Optional: Enhanced GUI features
matplotlib>=3.3.0  # For advanced plotting
scipy>=1.7.0       # For signal processing

# Optional: Video processing
# ffmpeg-python>=0.2.0  # For advanced video handling

# Documentation (optional)
# sphinx>=4.0.0
# sphinx-rtd-theme>=0.5.0
